diploma-chain backend/app/services/hedera_sdk_production.py
"""
DiplomaChain – Intégration Hedera SDK (Production)
═══════════════════════════════════════════════════

Ce fichier remplace hedera_service.py en production avec le vrai SDK.

INSTALLATION :
    pip install hedera-sdk-py

    OU via le wrapper Node.js (recommandé pour plus de stabilité) :
    npm install @hashgraph/sdk
    puis appeler via subprocess ou un microservice HTTP interne.

─────────────────────────────────────────────────────────────────
OPTION A : SDK Python natif (hedera-sdk-py)
─────────────────────────────────────────────────────────────────
"""
import json
import logging
from dataclasses import dataclass
from datetime import datetime, timezone

logger = logging.getLogger(__name__)


@dataclass
class HederaAnchorResult:
    success: bool
    transaction_id: str | None = None
    sequence_number: int | None = None
    consensus_timestamp: str | None = None
    error: str | None = None


class HederaSDKService:
    """
    Service Hedera utilisant le SDK Python officiel.
    
    Pour activer : remplacer les imports simulés par :
    
    from hedera import (
        Client, AccountId, PrivateKey, TopicId,
        TopicMessageSubmitTransaction,
        TopicCreateTransaction,
    )
    """

    def __init__(self, operator_id: str, operator_key: str, topic_id: str, network: str = "testnet"):
        self.operator_id = operator_id
        self.operator_key = operator_key
        self.topic_id = topic_id
        self.network = network
        self._client = None

    def _get_client(self):
        """Initialise le client Hedera (lazy loading)."""
        if self._client:
            return self._client

        # ── SDK Python (décommenter en production) ──────────────────────────
        # from hedera import Client, AccountId, PrivateKey
        #
        # if self.network == "mainnet":
        #     client = Client.for_mainnet()
        # else:
        #     client = Client.for_testnet()
        #
        # client.set_operator(
        #     AccountId.from_string(self.operator_id),
        #     PrivateKey.from_string(self.operator_key),
        # )
        # self._client = client
        # return client
        # ────────────────────────────────────────────────────────────────────

        raise NotImplementedError(
            "Installez hedera-sdk-py et décommentez le bloc SDK ci-dessus"
        )

    async def anchor_diploma(
        self,
        diploma_id: str,
        content_hash: str,
        institution_id: str,
        degree_title: str,
        graduation_date: str,
    ) -> HederaAnchorResult:
        """
        Soumet un message HCS avec le SDK Python officiel.
        
        Code SDK complet (à décommenter) :
        ────────────────────────────────────────────────────────────────────
        from hedera import (
            TopicId, TopicMessageSubmitTransaction
        )
        
        client = self._get_client()
        
        message_payload = json.dumps({
            "type": "DIPLOMA_ISSUANCE",
            "diploma_id": diploma_id,
            "content_hash": content_hash,
            "institution_id": institution_id,
            "degree_title": degree_title,
            "graduation_date": graduation_date,
            "issued_at": datetime.now(timezone.utc).isoformat(),
        })
        
        transaction = (
            TopicMessageSubmitTransaction()
            .set_topic_id(TopicId.from_string(self.topic_id))
            .set_message(message_payload)
        )
        
        # Exécuter et attendre le reçu
        response = transaction.execute(client)
        receipt = response.get_receipt(client)
        
        return HederaAnchorResult(
            success=True,
            transaction_id=str(response.transaction_id),
            sequence_number=receipt.topicSequenceNumber,
            consensus_timestamp=str(receipt.consensusTimestamp),
        )
        ────────────────────────────────────────────────────────────────────
        """
        raise NotImplementedError("SDK non initialisé")

    @staticmethod
    async def create_topic_sdk(operator_id: str, operator_key: str, memo: str = "DiplomaChain") -> str:
        """
        Crée un Topic HCS. À exécuter une seule fois au déploiement.
        
        Code SDK (à décommenter) :
        ────────────────────────────────────────────────────────────────────
        from hedera import (
            Client, AccountId, PrivateKey,
            TopicCreateTransaction
        )
        
        client = Client.for_testnet()  # ou for_mainnet()
        client.set_operator(
            AccountId.from_string(operator_id),
            PrivateKey.from_string(operator_key),
        )
        
        receipt = (
            TopicCreateTransaction()
            .set_topic_memo(memo)
            .execute(client)
            .get_receipt(client)
        )
        
        topic_id = str(receipt.topicId)
        print(f"✅ Topic créé : {topic_id}")
        print(f"Ajoutez dans .env : HEDERA_TOPIC_ID={topic_id}")
        return topic_id
        ────────────────────────────────────────────────────────────────────
        """
        raise NotImplementedError("SDK non initialisé")


# ─────────────────────────────────────────────────────────────────────────────
# OPTION B : Microservice Node.js (via HTTP interne)
# Plus robuste, SDK Node.js @hashgraph/sdk plus mature
# ─────────────────────────────────────────────────────────────────────────────
"""
server.js (microservice séparé) :

const { Client, AccountId, PrivateKey, TopicId, TopicMessageSubmitTransaction } = require("@hashgraph/sdk");
const express = require("express");
const app = express();
app.use(express.json());

const client = Client.forTestnet();
client.setOperator(
  AccountId.fromString(process.env.HEDERA_OPERATOR_ID),
  PrivateKey.fromString(process.env.HEDERA_OPERATOR_KEY)
);

app.post("/anchor", async (req, res) => {
  const { topic_id, message } = req.body;
  const response = await new TopicMessageSubmitTransaction()
    .setTopicId(TopicId.fromString(topic_id))
    .setMessage(JSON.stringify(message))
    .execute(client);
  const receipt = await response.getReceipt(client);
  res.json({
    transaction_id: response.transactionId.toString(),
    sequence_number: Number(receipt.topicSequenceNumber),
    consensus_timestamp: receipt.consensusTimestamp.toString(),
  });
});

app.listen(3001, () => console.log("Hedera microservice on :3001"));
"""

# Appel depuis Python :
# async with httpx.AsyncClient() as c:
#     r = await c.post("http://hedera-service:3001/anchor", json=payload)
#     data = r.json()
