diploma-chain backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
pymysql==1.1.1
cryptography==42.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
slowapi==0.1.9
python-dotenv==1.0.1
qrcode[pil]==7.4.2
Pillow==10.0.0
itsdangerous==2.2.0
email-validator==2.1.1
celery==5.4.0
redis==5.0.4
pypdf2==3.0.1
reportlab==4.1.0
pymupdf==1.24.0
